fix(ses)!: Withdraw support for muli-lockdown
*BREAKING CHANGE*: Lockdown can no longer be called more than once.  Lockdown no longer returns a boolean to indicate whether it was effective (true) or redundant (false).

Fixes #814
