+++ packages/ses-integration-test/test/test-sanity.js
@@ -3,7 +3,7 @@ import test from 'tape';
 import 'ses';
 
 test('sanity', t => {
-  t.equal(lockdown(), true, 'lockdown runs successfully');
+  t.equal(lockdown(), undefined, 'lockdown runs successfully');
   const c = new Compartment({ abc: 456 });
   t.equal(c.evaluate('123'), 123, 'simple evaluate succeeds');
   t.equal(c.evaluate('abc'), 456, 'endowment succeeds');
